# fix: Expose `globalObject` on `proxyHandler` [CEReactions] methods