--- a/lib/constructs/interface.js
+++ b/lib/constructs/interface.js
@@ -639,7 +639,7 @@ class Interface {
         `"Failed to set the " + index + " property on '${this.name}': The provided value"`);
       this.requires.merge(conv.requires);
 
-      const prolog = `
+      let prolog = `
         const index = ${P} >>> 0;
         let indexedValue = ${V};
         ${conv.body}
@@ -662,6 +662,8 @@ class Interface {
       }
 
       if (utils.hasCEReactions(this.indexedSetter)) {
+        prolog = `const globalObject = ${O}[impl]._globalObject;` + prolog;
+
         invocation = this.ctx.invokeProcessCEReactions(invocation, {
           requires: this.requires
         });
@@ -678,7 +680,7 @@ class Interface {
         `"Failed to set the '" + ${P} + "' property on '${this.name}': The provided value"`);
       this.requires.merge(conv.requires);
 
-      const prolog = `
+      let prolog = `
         let namedValue = ${V};
         ${conv.body}
       `;
@@ -701,6 +703,9 @@ class Interface {
       }
 
       if (utils.hasCEReactions(this.namedSetter)) {
+        prolog = `
+          const globalObject = ${O}[impl]._globalObject;` + prolog;
+
         invocation = this.ctx.invokeProcessCEReactions(invocation, {
           requires: this.requires
         });