--- a/lib/constructs/interface.js
+++ b/lib/constructs/interface.js
@@ -90,6 +90,7 @@ class Interface {
     this.namedSetter = null;
     this.namedDeleter = null;
     this.stringifier = null;
+    this.needsPerGlobalProxyHandler = false;
 
     this.iterable = null;
     this._analyzed = false;
@@ -229,11 +230,17 @@ class Interface {
               throw new Error(msg + "duplicated indexed setter");
             }
             this.indexedSetter = member;
+            if (utils.hasCEReactions(member)) {
+              this.needsPerGlobalProxyHandler = true;
+            }
           } else if (isNamed(member)) {
             if (this.namedSetter) {
               throw new Error(msg + "duplicated named setter");
             }
             this.namedSetter = member;
+            if (utils.hasCEReactions(member)) {
+              this.needsPerGlobalProxyHandler = true;
+            }
           } else {
             throw new Error(msg + "setter is neither indexed nor named");
           }
@@ -253,6 +260,9 @@ class Interface {
               throw new Error(msg + "duplicated named deleter");
             }
             this.namedDeleter = member;
+            if (utils.hasCEReactions(member)) {
+              this.needsPerGlobalProxyHandler = true;
+            }
           } else {
             throw new Error(msg + "deleter is not named");
           }
@@ -639,7 +649,7 @@ class Interface {
         `"Failed to set the " + index + " property on '${this.name}': The provided value"`);
       this.requires.merge(conv.requires);
 
-      let prolog = `
+      const prolog = `
         const index = ${P} >>> 0;
         let indexedValue = ${V};
         ${conv.body}
@@ -662,8 +672,6 @@ class Interface {
       }
 
       if (utils.hasCEReactions(this.indexedSetter)) {
-        prolog = `const globalObject = ${O}[impl]._globalObject;` + prolog;
-
         invocation = this.ctx.invokeProcessCEReactions(invocation, {
           requires: this.requires
         });
@@ -680,11 +688,10 @@ class Interface {
         `"Failed to set the '" + ${P} + "' property on '${this.name}': The provided value"`);
       this.requires.merge(conv.requires);
 
-      let prolog = `
+      const prolog = `
         let namedValue = ${V};
         ${conv.body}
       `;
-
 
       let invocation;
       if (!this.namedSetter.name) {
@@ -703,9 +710,6 @@ class Interface {
       }
 
       if (utils.hasCEReactions(this.namedSetter)) {
-        prolog = `
-          const globalObject = ${O}[impl]._globalObject;` + prolog;
-
         invocation = this.ctx.invokeProcessCEReactions(invocation, {
           requires: this.requires
         });
@@ -714,9 +718,21 @@ class Interface {
       return prolog + invocation;
     };
 
-    this.str += `
-      const proxyHandler = {
-    `;
+    let sep = "";
+    if (this.needsPerGlobalProxyHandler) {
+      this.str += `
+        const proxyHandlerCache = new WeakMap();
+        class ProxyHandler {
+          constructor(globalObject) {
+            this._globalObject = globalObject;
+          }
+      `;
+    } else {
+      this.str += `
+        const proxyHandler = {
+      `;
+      sep = ",";
+    }
 
     // [[Get]] (necessary because of proxy semantics)
     this.str += `
@@ -740,7 +756,7 @@ class Interface {
             return undefined;
           }
           return Reflect.apply(getter, receiver, []);
-        },
+        }${sep}
     `;
 
     // [[HasProperty]] (necessary because of proxy semantics)
@@ -758,7 +774,7 @@ class Interface {
             return Reflect.has(parent, P);
           }
           return false;
-        },
+        }${sep}
     `;
 
     // [[OwnPropertyKeys]]
@@ -789,7 +805,7 @@ class Interface {
             keys.add(key);
           }
           return [...keys];
-        },
+        }${sep}
     `;
 
     // [[GetOwnProperty]]
@@ -863,7 +879,7 @@ class Interface {
     }
     this.str += `
           return Reflect.getOwnPropertyDescriptor(target, P);
-        },
+        }${sep}
     `;
 
     // [[Set]]
@@ -874,6 +890,12 @@ class Interface {
           }
           if (target === receiver) {
     `;
+
+    if (this.needsPerGlobalProxyHandler) {
+      this.str += `
+          const globalObject = this._globalObject;
+      `;
+    }
 
     if (this.supportsIndexedProperties) {
       if (hasIndexedSetter) {
@@ -971,7 +993,7 @@ class Interface {
             valueDesc = { writable: true, enumerable: true, configurable: true, value: V };
           }
           return Reflect.defineProperty(receiver, P, valueDesc);
-        },
+        }${sep}
     `;
 
     // [[DefineOwnProperty]]
@@ -981,6 +1003,13 @@ class Interface {
             return Reflect.defineProperty(target, P, desc);
           }
     `;
+
+    if (this.needsPerGlobalProxyHandler) {
+      this.str += `
+          const globalObject = this._globalObject;
+      `;
+    }
+
     if (this.supportsIndexedProperties) {
       this.str += `
           if (utils.isArrayIndexPropName(P)) {
@@ -1059,7 +1088,7 @@ class Interface {
       `;
     }
     this.str += `
-        },
+        }${sep}
     `;
 
     // [[Delete]]
@@ -1069,6 +1098,13 @@ class Interface {
             return Reflect.deleteProperty(target, P);
           }
     `;
+
+    if (this.needsPerGlobalProxyHandler) {
+      this.str += `
+          const globalObject = this._globalObject;
+      `;
+    }
+
     if (this.supportsIndexedProperties) {
       this.str += `
           if (utils.isArrayIndexPropName(P)) {
@@ -1114,7 +1150,7 @@ class Interface {
     }
     this.str += `
           return Reflect.deleteProperty(target, P);
-        },
+        }${sep}
     `;
 
     // [[PreventExtensions]]
@@ -1173,9 +1209,22 @@ class Interface {
     `;
 
     if (this.isLegacyPlatformObj) {
-      this.str += `
+      if (this.needsPerGlobalProxyHandler) {
+        this.str += `
+        {
+          let proxyHandler = proxyHandlerCache.get(globalObject);
+          if (proxyHandler === undefined) {
+            proxyHandler = new ProxyHandler(globalObject);
+            proxyHandlerCache.set(globalObject, proxyHandler);
+          }
+          obj = new Proxy(obj, proxyHandler);
+        }
+        `;
+      } else {
+        this.str += `
         obj = new Proxy(obj, proxyHandler);
-      `;
+        `;
+      }
     }
 
     this.str += `