--- a/lib/constructs/interface.js
+++ b/lib/constructs/interface.js
@@ -1505,7 +1505,6 @@ class Interface {
 
     this.str += `
       exports.install = function install(globalObject) {
-        const interfaceName = "${name}";
     `;
 
     if (idl.inheritance) {
@@ -1540,6 +1539,9 @@ class Interface {
   }
 
   generate() {
+    this.str += `
+      const interfaceName = "${this.name}";
+    `;
     this.generateIterator();
 
     this.generateExport();