… test/__snapshots__/test.js.snap