Needed&nbsp;for&nbsp;<https://github.com/jsdom/jsdom/pull/2548>.

---

See&nbsp;also: <https://github.com/jsdom/webidl2js/pull/162>