@domenic 2020-02-01
Hmm, I'm probably missing something obvious, but why does this need to be done here instead of in the jsdom CEReactions processor code?

Also, the test snapshots aren't super-helpful in figuring this out, because they just declare a globalObject variable and then don't use it. Could we add a test that uses it, to illustrate when this is helpful?

Finally, I'll note that there's an alternate approach for getting per-instance state into proxy handlers, which is to make the proxy handler a class (so that methods are shared in the prototype) and then assign any per-instance state as fields on the class. We'd probably want to measure the performance of such an approach though.