@ExE-Boss 2020-02-02
The&nbsp;`globalObject` variable&nbsp;is&nbsp;used by&nbsp;the&nbsp;`CEReactions`&nbsp;steps:

https://github.com/jsdom/webidl2js/blob/246f79975a612fd15dff07844a6d9aec2a15135a/test/__snapshots__/test.js.snap#L505-L522