@domenic 2020-02-03
Was that an answer to my question of

> Hmm, I'm probably missing something obvious, but why does this need to be done here instead of in the jsdom CEReactions processor code?

or just my question about the test snapshots? If so, I don't quite understand.