@ExE-Boss 2020-02-03
After&nbsp;some&nbsp;refactoring, I’ve&nbsp;made&nbsp;it so&nbsp;that&nbsp;there’s now&nbsp;per‑global proxy&nbsp;handlers for&nbsp;interfaces that&nbsp;use&nbsp;`[CEReactions]` on&nbsp;named&nbsp;and&nbsp;indexed setters&nbsp;and&nbsp;deleters.

This&nbsp;keeps the&nbsp;performance&nbsp;gains from&nbsp;#156 for&nbsp;the&nbsp;most&nbsp;common&nbsp;case (no&nbsp;`[CEReactions]`), while&nbsp;ensuring that&nbsp;#130 doesn’t&nbsp;break.