@domenic 2020-02-03
Given that [CEReactions] is on the majority of classes in jsdom, and definitely the majority of performance-sensitive ones, I think we should benchmark such a change before committing to it.