@pmdartus 2020-02-10
I ran the performance benchmark in jsdom, and I was able to measure a 5-10% performance regressions when introducing this change on certain benchmarks. The benchmark is comparing the current master of jsdom against the custom elements branch + this PR. The performance improvements can probably be attributed to the create element refactor that has been done on the custom element branch.

IMO the performance regression is acceptable compared to value-added by custom element introduction (even if I am completely biased). I am counting a lot on https://github.com/jsdom/webidl2js/pull/158 to improve the overall performance even after the introduction of this regression.

<details>
<summary>Results</summary>

| Test name                                         | Master  (ops/sec) | CE + WeakMap (ops/sec) | Delta   |
|---------------------------------------------------|-------------------|------------------------|---------|
| dom/compare-document-position/compare ancestor    | 20,866            | 16,475                 | -21.04% |
| dom/compare-document-position/compare descendant  | 19,536            | 18,565                 | -4.97%  |
| dom/compare-document-position/compare siblings    | 901,133           | 898,123                | -0.33%  |
| dom/construction/createComment                    | 1,165,566         | 1,301,560              | 11.67%  |
| dom/construction/createDocumentFragment           | 1,320,087         | 1,331,535              | 0.87%   |
| dom/construction/createElement                    | 185,593           | 262,536                | 41.46%  |
| dom/construction/createEvent                      | 189,774           | 210,985                | 11.18%  |
| dom/construction/createNodeIterator               | 1,225,422         | 1,228,464              | 0.25%   |
| dom/construction/createProcessingInstruction      | 648,505           | 645,610                | -0.45%  |
| dom/construction/createTextNode                   | 1,008,861         | 1,037,554              | 2.84%   |
| dom/inner-html/tables                             | 0.31              | 0.29                   | -6.45%  |
| dom/named-properties/setAttribute                 | 3,110             | 3,238                  | 4.12%   |
| dom/tree-modification/appendChild: many parents   | 22.46             | 19.9                   | -11.40% |
| dom/tree-modification/appendChild: many siblings  | 246,965           | 238,447                | -3.45%  |
| dom/tree-modification/appendChild: no siblings    | 244,536           | 212,991                | -12.90% |
| dom/tree-modification/insertBefore: many siblings | 207,314           | 192,440                | -7.17%  |
| dom/tree-modification/removeChild: many parents   | 187,021           | 192,440                | 2.90%   |
| dom/tree-modification/removeChild: many siblings  | 127,226           | 142,754                | 12.21%  |
| dom/tree-modification/removeChild: no siblings    | 106,098           | 115,957                | 9.29%   |
| html/parsing/text                                 | 13.08             | 14.46                  | 10.55%  |
| jsdom/api/new JSDOM() defaults                    | 249               | 261                    | 4.82%   |
| jsdom/api/new JSDOM() with many elements          | 25.36             | 24.38                  | -3.86%  |
</details>