@domenic 2020-02-15
I went to merge this but I'm not sure I can write an accurate commit message. Would the following be an accurate summary of the changes?

```
Put globalObject into scope for processCEReactions replacement code

Now the replacement code returned by the processCEReactions handler can reference the globalObject variable, as is documented in the README. This was meant to be introduced in fda810e8d36fe64612bba533228b1b4e91f1a569, but the previous merging of 7c8037ffd4afcff01fd4ac48423675aa64aa1cea prevented that from working properly.
```

I wonder, is `interfaceName` (also documented in the readme) properly in scope still?