@ExE-Boss 2020-02-15
Since `interfaceName` is a constant shared across all globals, I’ve moved it into the enclosing module scope.